Connect to MongoDB
Connect Mongoose to MongoDB. Import Monggose to Server.js to invoke.
Encountered error while trying to connect to MongoDB.

rror:The `uri` parameter to `openUri()` must be a string, got "undefined". Make sure the first parameter to `mongoose.connect()` or `mongoose.createConnection()` is a string.
[nodemon] app crashed - waiting for file changes before starting..

It was working before without the Mongoose.